fix captcha

+++ src/main/java/me/towecraft/auth/listeners/captcha/CaptchaListener.java
@@ -50,6 +50,7 @@ public class CaptchaListener implements Listener {
                 if (e.getPlayer() != null && e.getPlayer() instanceof Player) {
                     Player player = (Player) e.getPlayer();
                     if (player.isOnline()) {
+                        captchaService.removeTypeCaptcha(player);
                         if (captchaService.getMapActions().get(e.getPlayer().getName()) != null &&
                                 captchaService.getMapActions().get(e.getPlayer().getName()).getCountDoneClick() < 3) {
                             if (captchaService.getTypeCaptcha(player) != TypeCaptcha.NONE) {
@@ -78,43 +79,39 @@ public class CaptchaListener implements Listener {
                     if (e.getSlot() >= 0) {
                         ItemStack item = e.getInventory().getItem(e.getSlot());
 
-                        if (item == null) {
+                        if (item != null &&
+                                !(e.getInventory().getItem(e.getSlot()).getType() == Material.STAINED_CLAY &&
+                                        e.getInventory().getItem(e.getSlot()).getType().getMaxDurability() == 14)) {
+                            ItemStack done = new ItemStack(Material.STAINED_CLAY, 1, (byte) 5);
+                            ItemMeta meta = done.getItemMeta();
+                            meta.setDisplayName(plugin.getConfig().getString("Captcha.nameSuccessItem",
+                                    "Not found String [Captcha.nameSuccessItem] in config.yml"));
+
+                            done.setItemMeta(meta);
+
+                            e.getInventory().setItem(e.getSlot(), done);
+                            captchaModel.setCountDoneClick(captchaModel.getCountDoneClick() + 1);
+                            captchaService.getMapActions().put(player.getName(), captchaModel);
+
+                            if (captchaModel.getCountDoneClick() >= captchaService.getCountDone()) {
+                                player.closeInventory();
+                                captchaService.removeTypeCaptcha(player);
+                                playerService.verify(player, false);
+                                return;
+                            }
+
+                            if (captchaService.getTypeCaptcha(player) == TypeCaptcha.SHOW_ONE_ITEM_HIDE_DONE ||
+                                    captchaService.getTypeCaptcha(player) == TypeCaptcha.SHOW_ONE_ITEM) {
+                                if (captchaService.getTypeCaptcha(player) == TypeCaptcha.SHOW_ONE_ITEM_HIDE_DONE) {
+                                    e.getInventory().setItem(e.getSlot(), null);
+                                }
+                                captchaService.nextShow(e);
+                            }
+                        } else {
                             captchaModel.setCountMissClick(captchaModel.getCountMissClick() + 1);
                             captchaService.getMapActions().put(player.getName(), captchaModel);
-                        } else {
-                            if (!e.getInventory().getItem(e.getSlot()).getItemMeta().getDisplayName()
-                                    .contains(plugin.getConfig().getString("Captcha.nameSuccessItem",
-                                            "Not found String [Captcha.nameSuccessItem] in config.yml"))) {
-                                ItemStack done = new ItemStack(Material.STAINED_CLAY, 1, (byte) 5);
-                                ItemMeta meta = done.getItemMeta();
-                                meta.setDisplayName(plugin.getConfig().getString("Captcha.nameSuccessItem",
-                                        "Not found String [Captcha.nameSuccessItem] in config.yml"));
-
-                                done.setItemMeta(meta);
-
-                                e.getInventory().setItem(e.getSlot(), done);
-                                captchaModel.setCountDoneClick(captchaModel.getCountDoneClick() + 1);
-                                captchaService.getMapActions().put(player.getName(), captchaModel);
-
-                                if (captchaModel.getCountDoneClick() >= captchaService.getCountDone()) {
-                                    player.closeInventory();
-                                    captchaService.removeTypeCaptcha(player);
-                                    playerService.verify(player, false);
-                                    return;
-                                }
-
-                                if (captchaService.getTypeCaptcha(player) == TypeCaptcha.SHOW_ONE_ITEM_HIDE_DONE ||
-                                        captchaService.getTypeCaptcha(player) == TypeCaptcha.SHOW_ONE_ITEM) {
-                                    if (captchaService.getTypeCaptcha(player) == TypeCaptcha.SHOW_ONE_ITEM_HIDE_DONE) {
-                                        e.getInventory().setItem(e.getSlot(), null);
-                                    }
-                                    captchaService.nextShow(e);
-                                }
-                            } else {
-                                captchaModel.setCountMissClick(captchaModel.getCountMissClick() + 1);
-                                captchaService.getMapActions().put(player.getName(), captchaModel);
-                            }
                         }
+
                         if (captchaModel.getCountMissClick() > captchaService.getCountMiss()) {
                             printMessage.kickMessage(player, fileMessages.getMSG().getString("KickMessages.youBot",
                                     "Not found string [KickMessages.youBot]"));
